workaround rb bug when downloading sources
