+++ packit/base_git.py
@@ -32,7 +32,7 @@ from packit.actions import ActionName
 from packit.config import Config, PackageConfig
 from packit.exceptions import PackitException
 from packit.local_project import LocalProject
-from packit.utils import run_command
+from packit.utils import run_command, cwd
 
 logger = logging.getLogger(__name__)
 
@@ -295,4 +295,17 @@ class PackitRepositoryBase:
         return None
 
     def fetch_upstream_archive(self):
-        self.specfile.download_remote_sources()
+        # TODO: there is a bug in rebase-helper that it saves the source in cwd
+        #   File "/src/packit/base_git.py", line 298, in fetch_upstream_archive
+        #     self.specfile.download_remote_sources()
+        #   File "rebasehelper/specfile.py", line 220, in download_remote_sources
+        #     LookasideCacheHelper.download('fedpkg', os.path.dirname(self.path), self.get_
+        #   File "rebasehelper/helpers/lookaside_cache_helper.py", line 122, in download
+        #     cls._download_source(tool, url, package, source['filename'], source['hashtype'],
+        #   File "rebasehelper/helpers/lookaside_cache_helper.py", line 109, in _download_source
+        #     try:
+        #   File "rebasehelper/helpers/download_helper.py", line 162, in download_file
+        #     with open(destination_path, 'wb') as local_file:
+        # PermissionError: [Errno 13] Permission denied: 'systemd-8bca462.tar.gz'
+        with cwd(self.specfile_dir):
+            self.specfile.download_remote_sources()
