+++ tests/integration/test_build.py
@@ -24,7 +24,7 @@ from packit.api import PackitAPI
 from packit.config import get_local_package_config
 from packit.local_project import LocalProject
 from tests.spellbook import get_test_config
-from tests.utils import cwd
+from packit.utils import cwd
 
 
 def test_basic_build(upstream_n_distgit, mock_remote_functionality_upstream):
