+++ tests/integration/test_create_update.py
@@ -28,7 +28,7 @@ from packit.api import PackitAPI
 from packit.config import get_local_package_config
 from packit.local_project import LocalProject
 from tests.spellbook import get_test_config, can_a_module_be_imported
-from tests.utils import cwd
+from packit.utils import cwd
 
 
 @pytest.fixture()
